+++ espresso-macchiato/src/androidTest/java/de/nenick/espressomacchiato/elements/EspViewTest.java
@@ -1,6 +1,8 @@
 package de.nenick.espressomacchiato.elements;
 
 import android.support.test.espresso.NoMatchingViewException;
+import android.view.GestureDetector;
+import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ScrollView;
@@ -17,6 +19,7 @@ import de.nenick.espressomacchiato.testbase.EspressoTestCase;
 public class EspViewTest extends EspressoTestCase<BaseActivity> {
 
     private static final String VIEW_WAS_CLICKED_MESSAGE = "view was clicked";
+    private static final String VIEW_WAS_DOUBLE_CLICKED_MESSAGE = "view was double clicked";
 
     private int viewId = android.R.id.button1;
     private EspView espView = EspView.byId(viewId);
@@ -68,19 +71,6 @@ public class EspViewTest extends EspressoTestCase<BaseActivity> {
         espView.assertIsDisplayedOnScreen();
     }
 
-    private void givenViewLargerThanScreen() {
-        ScrollView scrollView = new ScrollView(getActivity());
-        scrollView.setId(android.R.id.candidatesArea);
-        addViewToLayout(scrollView, BaseActivity.rootLayout);
-
-        view = new Button(activityTestRule.getActivity());
-        view.setHeight(3000);
-        view.setWidth(2000);
-        view.setId(viewId);
-        addViewToLayout(view, android.R.id.candidatesArea);
-    }
-
-
     @Test
     public void testAssertIsDisplayedOnScreenFailure() {
         givenViewOutsideOfScreen(android.R.id.text2);
@@ -89,6 +79,7 @@ public class EspViewTest extends EspressoTestCase<BaseActivity> {
         exception.expect(AssertionFailedError.class);
         EspView.byId(android.R.id.text2).assertIsDisplayedOnScreen();
     }
+
 
     @Test
     public void testClick() {
@@ -128,6 +119,29 @@ public class EspViewTest extends EspressoTestCase<BaseActivity> {
         MyEspView myEspView = new MyEspView(EspView.byId(viewId));
         myEspView.click();
         espTextView.assertTextIs(VIEW_WAS_CLICKED_MESSAGE);
+    }
+
+    @Test
+    public void testDoubleClick() {
+        givenDoubleClickableView();
+        givenClickFeedbackTextView();
+
+        espView.doubleClick();
+
+        espTextView.assertTextIs(VIEW_WAS_DOUBLE_CLICKED_MESSAGE);
+    }
+
+
+    private void givenViewLargerThanScreen() {
+        ScrollView scrollView = new ScrollView(getActivity());
+        scrollView.setId(android.R.id.candidatesArea);
+        addViewToLayout(scrollView, BaseActivity.rootLayout);
+
+        view = new Button(activityTestRule.getActivity());
+        view.setHeight(3000);
+        view.setWidth(2000);
+        view.setId(viewId);
+        addViewToLayout(view, android.R.id.candidatesArea);
     }
 
     private void givenViewIsInvisible() {
@@ -170,6 +184,49 @@ public class EspViewTest extends EspressoTestCase<BaseActivity> {
         });
     }
 
+    private void givenDoubleClickableView() {
+        view = new Button(activityTestRule.getActivity());
+        view.setId(viewId);
+        addViewToLayout(view, BaseActivity.rootLayout);
+
+        view.setOnTouchListener(new View.OnTouchListener() {
+            class MySimpleOnGestureListener extends GestureDetector.SimpleOnGestureListener {
+                @Override
+                public boolean onDoubleTap(MotionEvent e) {
+                    messageView.setText(VIEW_WAS_DOUBLE_CLICKED_MESSAGE);
+                    return super.onDoubleTap(e);
+                }
+            }
+
+            GestureDetector gestureDetector = null;
+            MySimpleOnGestureListener mySimpleOnGestureListener;
+            long lastClick;
+
+            @Override
+            public boolean onTouch(View v, MotionEvent event) {
+                if (gestureDetector == null) {
+                    mySimpleOnGestureListener = new MySimpleOnGestureListener();
+                    gestureDetector = new GestureDetector(activityTestRule.getActivity(), mySimpleOnGestureListener);
+                }
+                if (gestureDetector.onTouchEvent(event)) {
+                    return true;
+                }
+
+                // On slow devices (e.g. emulator) we need more delay to detect the double click.
+                // Delay from SimpleOnGestureListener is to less to detect it.
+                long currentTime = System.currentTimeMillis();
+                if (currentTime - lastClick < 1000) {
+                    mySimpleOnGestureListener.onDoubleTap(event);
+                    return true;
+                }
+                lastClick = currentTime;
+
+                return false;
+            }
+        });
+    }
+
+
     private void givenClickableViewNotOnScreen() {
         for (int i=0;i<20;i++) {
             addViewToLayout(new Button(activityTestRule.getActivity()), BaseActivity.rootLayout);
@@ -195,7 +252,7 @@ public class EspViewTest extends EspressoTestCase<BaseActivity> {
 
     /** Element extension */
     class MyEspView extends EspView {
-        public MyEspView(EspView template) {
+        MyEspView(EspView template) {
             super(template);
         }
     }